Separates eigen exceptions.
