+++ src/linear.cpp
@@ -1,6 +1,7 @@
 #include "linear.hpp"
 #include "exceptions/activation_functions.hpp"
 #include "exceptions/differentiable.hpp"
+#include "exceptions/eigen.hpp"
 #include "utils/exceptions.hpp"
 #include "utils/matrix.hpp"
 
@@ -42,7 +43,7 @@ Eigen::MatrixXd Linear::getWeight() const { return this->weight; }
 void Linear::setWeight(Eigen::MatrixXd weight) {
   if (this->weight.rows() != weight.rows() ||
       this->weight.cols() != weight.cols()) {
-    throw exceptions::eigen::InvalidShapeException();
+    throw exceptions::eigen::InvalidShapeException(this->weight, weight);
   }
   this->weight = weight;
 }
@@ -53,7 +54,7 @@ Eigen::MatrixXd Linear::getBias() const { return this->bias; };
 
 void Linear::setBias(Eigen::MatrixXd bias) {
   if (this->bias.rows() != bias.rows() || this->bias.cols() != bias.cols()) {
-    throw exceptions::eigen::InvalidShapeException();
+    throw exceptions::eigen::InvalidShapeException(this->bias, bias);
   }
   this->bias = bias;
 }
