+++ test/test_cross_entropy_loss.cpp
@@ -1,5 +1,6 @@
 #include "cross_entropy_loss.hpp"
 #include "exceptions/differentiable.hpp"
+#include "exceptions/eigen.hpp"
 #include "utils/exceptions.hpp"
 #include <gtest/gtest.h>
