+++ test/test_linear.cpp
@@ -1,5 +1,6 @@
 #include "exceptions/activation_functions.hpp"
 #include "exceptions/differentiable.hpp"
+#include "exceptions/eigen.hpp"
 #include "linear.hpp"
 #include "utils/exceptions.hpp"
 #include <Eigen/Dense>
